fix(host): keep Beads lifecycle fail-fast
Limit shared-database restore to explicit missing-database failures and stop re-running init for existing attachments that remain unready after repair. Reorganize host-infra-beads tests into lifecycle, parsing, and persistence modules so the extracted boundaries stay reviewable.
